PhyDNet_Add_On/requirement.txt
certifi==2020.12.5
chardet==4.0.0
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
cycler==0.10.0
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1615055117017/work
decorator @ file:///home/ktietz/src/ci/decorator_1611930055503/work
filelock==3.0.12
gdown==3.13.0
idna==2.10
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
natsort @ file:///tmp/build/80754af9/natsort_1611692802956/work
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
olefile==0.46
opencv-python==4.5.1.48
pandas==1.2.3
Pillow @ file:///tmp/build/80754af9/pillow_1615224089586/work
piqa==1.1.3
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks==1.7.1
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658317819/work
PyYAML==5.4.1
requests==2.25.1
scikit-image==0.17.2
scipy @ file:///tmp/build/80754af9/scipy_1614022789989/work
six @ file:///tmp/build/80754af9/six_1605205327372/work
tifffile==2020.10.1
toolz @ file:///home/linux1/recipes/ci/toolz_1610987900194/work
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchvision==0.9.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3==1.26.5
